WIP Conditionally ask for other parent's email
