+++ libs/application/templates/parental-leave/src/fields/Review/index.tsx
@@ -30,10 +30,9 @@ import {
 // import { getEstimatedPayments } from '../PaymentSchedule/estimatedPaymentsQuery'
 import { parentalLeaveFormMessages } from '../../lib/messages'
 import { YES, NO, MANUAL } from '../../constants'
-import { Period } from '../../types'
+import { Boolean } from '../../types'
 import { SummaryTimeline } from '../components/Timeline/SummaryTimeline'
 import { SummaryRights } from '../Rights/SummaryRights'
-import { Boolean } from '../../hooks/useApplicationAnswers'
 import { useUnion as useUnionOptions } from '../../hooks/useUnion'
 import { usePrivatePensionFund as usePrivatePensionFundOptions } from '../../hooks/usePrivatePensionFund'
 import { usePensionFund as usePensionFundOptions } from '../../hooks/usePensionFund'
