+++ libs/application/templates/parental-leave/src/hooks/useApplicationAnswers.ts
@@ -1,126 +1,9 @@
-import { Application, getValueViaPath } from '@island.is/application/core'
+import { Application } from '@island.is/application/core'
 
-import { MANUAL, NO, YES } from '../constants'
-
-type OtherParent = typeof NO | typeof MANUAL | undefined
-export type ApplicationAnswers = ReturnType<typeof useApplicationAnswers>
-export type Boolean = typeof NO | typeof YES
+import { getApplicationAnswers } from '../parentalLeaveUtils'
 
 export const useApplicationAnswers = (application: Application) => {
-  const answers = application.answers
-
-  const otherParent = getValueViaPath(answers, 'otherParent') as OtherParent
-
-  const pensionFund = getValueViaPath(answers, 'payments.pensionFund') as string
-
-  const union = getValueViaPath(answers, 'payments.union') as string
-
-  const usePrivatePensionFund = getValueViaPath(
-    answers,
-    'usePrivatePensionFund',
-  ) as Boolean
-
-  const privatePensionFund = getValueViaPath(
-    answers,
-    'payments.privatePensionFund',
-  ) as string
-
-  const privatePensionFundPercentage = getValueViaPath(
-    answers,
-    'payments.privatePensionFundPercentage',
-  ) as string
-
-  const isSelfEmployed = getValueViaPath(
-    answers,
-    'employer.isSelfEmployed',
-  ) as Boolean
-
-  const otherParentName = getValueViaPath(answers, 'otherParentName') as string
-
-  const otherParentId = getValueViaPath(answers, 'otherParentId') as string
-
-  const bank = getValueViaPath(answers, 'payments.bank') as string
-
-  const personalAllowance = getValueViaPath(
-    answers,
-    'usePersonalAllowance',
-  ) as Boolean
-
-  const personalAllowanceFromSpouse = getValueViaPath(
-    answers,
-    'usePersonalAllowanceFromSpouse',
-    NO,
-  ) as Boolean
-
-  const personalUseAsMuchAsPossible = getValueViaPath(
-    answers,
-    'personalAllowance.useAsMuchAsPossible',
-  ) as Boolean
-
-  const personalUsage = getValueViaPath(
-    answers,
-    'personalAllowance.usage',
-  ) as string
-
-  const spouseUseAsMuchAsPossible = getValueViaPath(
-    answers,
-    'personalAllowanceFromSpouse.useAsMuchAsPossible',
-  ) as Boolean
-
-  const spouseUsage = getValueViaPath(
-    answers,
-    'personalAllowanceFromSpouse.usage',
-  ) as string
-
-  const employerEmail = getValueViaPath(answers, 'employer.email') as string
-
-  const shareInformationWithOtherParent = getValueViaPath(
-    answers,
-    'shareInformationWithOtherParent',
-  ) as Boolean
-
-  const selectedChild = getValueViaPath(answers, 'selectedChild') as string
-
-  const isRequestingRights = getValueViaPath(
-    answers,
-    'requestRights.isRequestingRights',
-  ) as Boolean
-
-  const requestDays = getValueViaPath(
-    answers,
-    'requestRights.requestDays',
-  ) as number
-
-  const isGivingRights = getValueViaPath(
-    answers,
-    'giveRights.isGivingRights',
-  ) as Boolean
-
-  const giveDays = getValueViaPath(answers, 'giveRights.giveDays') as number
-
-  return {
-    otherParent,
-    pensionFund,
-    union,
-    usePrivatePensionFund,
-    privatePensionFund,
-    privatePensionFundPercentage,
-    isSelfEmployed,
-    otherParentName,
-    otherParentId,
-    bank,
-    personalAllowance,
-    personalAllowanceFromSpouse,
-    personalUseAsMuchAsPossible,
-    personalUsage,
-    spouseUseAsMuchAsPossible,
-    spouseUsage,
-    employerEmail,
-    shareInformationWithOtherParent,
-    selectedChild,
-    isRequestingRights,
-    requestDays,
-    isGivingRights,
-    giveDays,
-  }
+  return getApplicationAnswers(application.answers)
 }
+
+export type ApplicationAnswers = ReturnType<typeof useApplicationAnswers>
