+++ libs/application/templates/parental-leave/src/lib/dataSchema.ts
@@ -69,6 +69,7 @@ export const dataSchema = z.object({
       params: errorMessages.otherParentId,
     }),
   otherParentRightOfAccess: z.enum([YES, NO]).optional(),
+  otherParentEmail: z.string().email(),
   usePersonalAllowance: z.enum([YES, NO]),
   usePersonalAllowanceFromSpouse: z.enum([YES, NO]),
 })
