+++ libs/application/templates/parental-leave/src/types.ts
@@ -1,3 +1,5 @@
+import { YES, NO } from './constants'
+
 export type Period = {
   startDate: string
   endDate: string
@@ -20,3 +22,5 @@ export interface Payment {
     paid: boolean
   }
 }
+
+export type Boolean = typeof NO | typeof YES
